- (keytools.c, keytools.h, parse.c, scapi.c): Patch from Markku Laukkanen:   - size_t changes.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@2885 06827809-a52a-0410-b366-d66718629ded

+++ net-snmp/snmplib/keytools.c
@@ -41,7 +41,7 @@
 int (*kmt_hash) (
 	const int	  mode,		void  	 **context,
 	const u_int8_t	 *data,		const int  data_len,     
-	u_int8_t	**digest,	int	  *digest_len);
+	u_int8_t	**digest,	size_t	  *digest_len);
 
 /*******************************************************************-o-******
  * generate_Ku
